Add JwtTokenCode and JWT Filter code

+++ backend/src/main/java/com/example/backend/service/UserService.java
@@ -2,15 +2,17 @@ package com.example.backend.service;
 
 import com.example.backend.DTO.requestDTO.AddUserRequestDTO;
 import com.example.backend.DTO.requestDTO.EditUserRequestDTO;
+import com.example.backend.DTO.requestDTO.LoginRequestDTO;
 import com.example.backend.DTO.responseDTO.UserResponseDTO;
 import com.example.backend.entity.UserEntity;
 import com.example.backend.exception.ElementNotFound;
 import com.example.backend.repository.UserRepository;
 import com.example.backend.utils.Transformer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -28,7 +30,7 @@ public class UserService {
     }
 
     public UserResponseDTO getUserById(Long id){
-        UserEntity user =  userRepository.findById(id).orElseThrow(()->new ElementNotFound("User Not Found with id: " +id));
+        UserEntity user =  userRepository.findById(id).orElseThrow(); //ByDefault NoSuchElementException is given by orElseThrow()
         return Transformer.UserEntitytoUserResponseDTO(user);
     }
 
@@ -51,8 +53,20 @@ public class UserService {
         throw new ElementNotFound("No Such user found with this id:"+userId);
     }
 
-    public void deleteUserDetails(Long Id){
+    public void deleteUserById(Long Id){
+        if(!userRepository.existsById(Id)){
+            throw new UsernameNotFoundException("User not found");
+        }
         userRepository.deleteById(Id);
+    }
+
+    public void deleteAllUsers(){
+        userRepository.deleteAll();
+    }
+
+    public UserResponseDTO authenticateUser(@RequestBody LoginRequestDTO loginRequestDTO) {
+        UserEntity authUserEntity =  userRepository.findByUsername(loginRequestDTO.getUsername()).orElseThrow(()->new UsernameNotFoundException("user not found"));
+        return Transformer.UserEntitytoUserResponseDTO(authUserEntity);
     }
 
 }
